refactor how we initialize for higher scale on implicit-feedback
